py/08_EntornosVirtuales/01_venv/servidor/requirements.txt
blinker==1.9.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Flask==3.1.2
idna==3.4
itsdangerous==2.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
PyJWT==2.10.1
requests==2.28.1
urllib3==1.26.12
Werkzeug==3.1.3
